Bring back --default to allow OBS builds to select an entry point among many
When building a mkosi repository called 'particleos', OBS calls mkosi as such:

mkosi --directory /usr/src/packages/SOURCES/particleos --default mkosi.conf

with mkosi.conf being the source file extracted at source time, ie: defined
by the user, among the many files available in the repository.
So this can be used to select the entry point, that defines what to build,
when a repository provides many build types, profiles, images, etc.

Right now --default is a no-op, so make it add the passed conf file as a
high priority config file, so that the entry point can have precedence
over anything else.

The file is placed at:

/usr/src/packages/SOURCES/mkosi.conf

This cannot be controlled by the user, and it's fixed.

Config file parsing requires absolute paths, and aborts otherwise.
Because it is called with a filename only, but mkosi executes from the
--directory location, special case it and if a non-absolute path is passed,
look in the parent of --directory so that the use case of providing an
entry point can be enabled.